finish show data in admin managerment

+++ database/migrations/2024_09_29_135416_create_order_items_table.php
@@ -12,7 +12,7 @@ class CreateOrderItemsTable extends Migration
             $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
             $table->integer('quantity');
-            $table->decimal('price', 10, 2); // giá sản phẩm tại thời điểm đặt hàng (không phải giá hiện tại)
+            $table->integer('price'); // giá sản phẩm tại thời điểm đặt hàng (không phải giá hiện tại)
             $table->timestamps();
         });
     }
